+++ include/souper/Tool/CandidateMapUtils.h
@@ -17,7 +17,8 @@
 
 #include "llvm/Support/raw_ostream.h"
 #include "souper/Extractor/Candidates.h"
-#include "souper/Extractor/KLEEBuilder.h"
+//#include "souper/Extractor/KLEEBuilder.h"
+#include "souper/Extractor/ExprBuilder.h"
 #include "souper/Extractor/Solver.h"
 #include "souper/KVStore/KVStore.h"
