+++ unittests/Extractor/ExtractorTests.cpp
@@ -20,7 +20,8 @@
 #include "llvm/IR/Module.h"
 #include "llvm/Support/SourceMgr.h"
 #include "souper/Extractor/Candidates.h"
-#include "souper/Extractor/KLEEBuilder.h"
+//#include "souper/Extractor/KLEEBuilder.h"
+#include "souper/Extractor/ExprBuilder.h"
 #include <memory>
 #include "gtest/gtest.h"
